# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StandardTestItemAliasCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "alias_text": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Alias Text",
            "type": "string"
          },
          "material_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Id"
          },
          "model_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "standard_test_item_id": {
            "format": "uuid",
            "title": "Standard Test Item Id",
            "type": "string"
          },
          "supplier_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          }
        },
        "required": [
          "standard_test_item_id",
          "alias_text",
          "priority"
        ],
        "title": "StandardTestItemAliasCreateRequest",
        "type": "object"
      },
      "StandardTestItemAliasResponse": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "alias_text": {
            "title": "Alias Text",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lock_version": {
            "title": "Lock Version",
            "type": "integer"
          },
          "material_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Id"
          },
          "model_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "standard_test_item_id": {
            "format": "uuid",
            "title": "Standard Test Item Id",
            "type": "string"
          },
          "supplier_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "standard_test_item_id",
          "alias_text",
          "supplier_id",
          "material_id",
          "model_id",
          "priority",
          "active",
          "lock_version",
          "created_at",
          "updated_at"
        ],
        "title": "StandardTestItemAliasResponse",
        "type": "object"
      },
      "StandardTestItemAliasUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "alias_text": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Alias Text",
            "type": "string"
          },
          "material_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Id"
          },
          "model_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "standard_test_item_id": {
            "format": "uuid",
            "title": "Standard Test Item Id",
            "type": "string"
          },
          "supplier_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          }
        },
        "required": [
          "standard_test_item_id",
          "alias_text",
          "priority"
        ],
        "title": "StandardTestItemAliasUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "HYC Inspection API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/standard-test-item-aliases": {
      "get": {
        "operationId": "list_aliases_api_v1_standard_test_item_aliases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StandardTestItemAliasResponse"
                  },
                  "title": "Response List Aliases Api V1 Standard Test Item Aliases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Aliases",
        "tags": [
          "p5-aliases"
        ]
      },
      "post": {
        "operationId": "create_alias_api_v1_standard_test_item_aliases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardTestItemAliasCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardTestItemAliasResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Alias",
        "tags": [
          "p5-aliases"
        ]
      }
    },
    "/api/v1/standard-test-item-aliases/lookup": {
      "get": {
        "operationId": "lookup_aliases_api_v1_standard_test_item_aliases_lookup_get",
        "parameters": [
          {
            "in": "query",
            "name": "alias_text",
            "required": true,
            "schema": {
              "title": "Alias Text",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "supplier_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Supplier Id"
            }
          },
          {
            "in": "query",
            "name": "material_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Material Id"
            }
          },
          {
            "in": "query",
            "name": "model_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StandardTestItemAliasResponse"
                  },
                  "title": "Response Lookup Aliases Api V1 Standard Test Item Aliases Lookup Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lookup Aliases",
        "tags": [
          "p5-aliases"
        ]
      }
    },
    "/api/v1/standard-test-item-aliases/{alias_id}": {
      "delete": {
        "operationId": "delete_alias_api_v1_standard_test_item_aliases__alias_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "alias_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alias Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Alias",
        "tags": [
          "p5-aliases"
        ]
      },
      "get": {
        "operationId": "get_alias_api_v1_standard_test_item_aliases__alias_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "alias_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alias Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardTestItemAliasResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Alias",
        "tags": [
          "p5-aliases"
        ]
      },
      "put": {
        "operationId": "update_alias_api_v1_standard_test_item_aliases__alias_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "alias_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alias Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardTestItemAliasUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardTestItemAliasResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Alias",
        "tags": [
          "p5-aliases"
        ]
      }
    },
...
  }
}